Configured form data processing and saving to the database with dynamic data submission via JavaScript without page reload.
PHP functions for data processing:

empty Value(): Checks for the presence of values in the array and retrieves them.
PFC(): Calculation and formatting of the ratio of proteins, fats and carbohydrates.
zero(): Formatting numeric values for display.
insert(): Inserts data into a database with preparation and execution of an SQL query.
The main PHP script:

Connecting the necessary components.
Processing data from a POST request.
Uploading an image of the dish.
Calculation and filling of additional data (PFC).
Inserting data into the database.
JavaScript for submitting the form:

Adding a handler for the form submission event.
Sending form data via an AJAX request.
Processing the server response and output the result to the console.